Add wallet menu item

+++ packages/neuron-wallet/src/controllers/wallets/index.ts
@@ -1,3 +1,5 @@
+import fs from 'fs'
+import AppController from '../app'
 import WalletsService, { Wallet, WalletProperties } from '../../services/wallets'
 import Key from '../../keys/key'
 import { Controller as ControllerDecorator, CatchControllerError } from '../../decorators'
@@ -7,8 +9,11 @@ import {
   IsRequired,
   WalletNotFound,
   IncorrectPassword,
+  EmptyPassword,
   ServiceHasNoResponse,
 } from '../../exceptions'
+import prompt from '../../utils/prompt'
+import i18n from '../../utils/i18n'
 
 const walletsService = WalletsService.getInstance()
 
@@ -161,9 +166,16 @@ export default class WalletsController {
   }
 
   @CatchControllerError
-  public static async delete({ id, password }: { id: string; password: string }): Promise<Controller.Response<any>> {
-    if (!walletsService.validate({ id, password })) throw new IncorrectPassword()
+  public static async delete(id: string): Promise<Controller.Response<any>> {
+    const password = await WalletsController.requestPassword(i18n.t('messageBox.remove-wallet.title'))
+    if (password === null)
+      return {
+        status: ResponseCode.Success,
+        result: null,
+      }
 
+    if (password === '') throw new EmptyPassword()
+    if (!walletsService.validate({ id, password })) throw new IncorrectPassword()
     walletsService.delete(id)
 
     return {
@@ -182,6 +194,17 @@ export default class WalletsController {
       status: ResponseCode.Success,
       result: JSON.stringify(walletsService.get(id)),
     }
+  }
+
+  @CatchControllerError
+  public static async backup(id: string): Promise<Controller.Response<boolean>> {
+    const password = await WalletsController.requestPassword(i18n.t('messageBox.backup-keystore.title'))
+    if (password === null)
+      return {
+        status: ResponseCode.Success,
+        result: false,
+      }
+    return WalletsController.downloadKeystore(id, password)
   }
 
   @CatchControllerError
@@ -240,6 +263,39 @@ export default class WalletsController {
         },
       }
     }
+  }
+
+  private static async requestPassword(title: string): Promise<string | null> {
+    const password = (await prompt('password', {
+      title,
+    })) as string | null
+    return password
+  }
+
+  private static async downloadKeystore(id: string, password: string): Promise<Controller.Response<boolean>> {
+    if (password === '') throw new EmptyPassword()
+    const wallet = await walletsService.get(id)
+
+    if (!walletsService.validate({ id, password })) throw new IncorrectPassword()
+
+    const keystore = wallet.loadKeystore()
+    return new Promise(resolve => {
+      AppController.showSaveDialog(
+        {
+          title: i18n.t('messages.save-keystore'),
+          defaultPath: wallet.name,
+        },
+        (filename?: string) => {
+          if (filename) {
+            fs.writeFileSync(filename, JSON.stringify(keystore))
+            resolve({
+              status: ResponseCode.Success,
+              result: true,
+            })
+          }
+        }
+      )
+    })
   }
 }
